requirements.txt
python-dotenv==0.13.0
python-telegram-bot==12.7
requests==2.23.0